+++ Login/src/main/java/com/portal/service/GroupService.java
@@ -1,9 +1,9 @@
 package com.portal.service;
 
-import com.portal.model.Group;
+import com.portal.entity.Group;
 
 public interface GroupService {
 	
-	public Group getGroup(int id);
+	public Group getGroup(Long id);
 
 }
