+++ Login/src/main/java/com/portal/service/GroupServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.portal.dao.GroupDAO;
-import com.portal.model.Group;
+import com.portal.entity.Group;
 
 @Service
 @Transactional
@@ -14,7 +14,7 @@ public class GroupServiceImpl implements GroupService {
 	@Autowired
 	private GroupDAO groupDAO;
 
-	public Group getGroup(int id) {
+	public Group getGroup(Long id) {
 		return groupDAO.getGroup(id);
 	}
